requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
requests>=2.28.0
pyyaml>=6.0
py-solc-x>=2.0.0
